Refactor duplicate queue.add() calls into one.
